% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{adductData}
\alias{adductData}
\alias{adductData-package}
\title{Adduct data from Grigoryan at al 2016.
This package contains 2 mzXML files from Grigoryan et al
that were used to perform adductomics.}
\source{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC5555296/}
}
\value{
2 mzXML files and processed data
}
\description{
Adduct data from Grigoryan at al 2016.
This package contains 2 mzXML files from Grigoryan et al
that were used to perform adductomics.
}
\examples{
\dontrun{
hub <- ExperimentHub::ExperimentHub()
x <- AnnotationHub::query(hub, c("adductData"))
## download resource
data = x[[1]]
}}
\keyword{utilities}



























